project/requirements.txt
asgiref==3.2.5
Django==3.0.7
djangorestframework==3.12.1
python-dotenv==0.12.0
pytz==2019.3
sqlparse==0.3.1
